Whitelist email attribute for mass assignment
ActiveRecord whitelist_attributes is set to true by default as of Rails 3.2.3. Any attributes that should be mass assignable need to be enabled by attr_accessor. This change makes that so on prelaunch_subscriber so that one does not have to disable whitelist_attributes which is an important security feature.